+++ R/select.R
@@ -2,13 +2,14 @@
 #'
 #' @param conn A DBIConnection object
 #' @param statement The SQL statement to use
+#' @param retries Max retries if query fails
 #' @export
 #' @examples
 #' db <- dbxConnect(adapter="sqlite", dbname=":memory:")
 #' DBI::dbCreateTable(db, "forecasts", data.frame(id=1:3, temperature=20:22))
 #'
 #' records <- dbxSelect(db, "SELECT * FROM forecasts")
-dbxSelect <- function(conn, statement) {
+dbxSelect <- function(conn, statement, retries=0) {
   statement <- processStatement(statement)
   cast_dates <- list()
   cast_datetimes <- list()
@@ -18,7 +19,7 @@ dbxSelect <- function(conn, statement) {
   unescape_blobs <- list()
   fix_timetz <- list()
 
-  r <- fetchRecords(conn, statement)
+  r <- withRetries(retries, function() { fetchRecords(conn, statement) })
   records <- r$records
   column_info <- r$column_info
 
@@ -155,6 +156,20 @@ fetchRecords <- function(conn, statement) {
   })
 
   list(records=combineResults(ret), column_info=column_info)
+}
+
+withRetries <- function(retries, fn) {
+  err <- NULL
+
+  for (i in 1:(retries + 1)) {
+    tryCatch({
+      return(fn())
+    }, error=function(e) {
+      err <<- e
+    })
+  }
+
+  stop(err)
 }
 
 emptyType <- function(type) {